Fix CLI app compatibility
