+++ Project.toml
@@ -32,6 +32,6 @@ KomaMRIFiles = "0.10"
 KomaMRIPlots = "0.11"
 MAT = "0.10, 0.11, 0.12"
 MRIReco = "0.6, 0.7, 0.8, 0.9"
-PrecompileTools = "1.3.3"
+PrecompileTools = "1"
 Reexport = "1"
 julia = "1.9"
