+++ src/KomaCLI.jl
@@ -276,23 +276,3 @@ end
         end
     end
 end
-
-module CLI
-
-import ..KomaMRI: CLI_HELP, run_cli
-
-function print_help()
-    print(CLI_HELP)
-    return nothing
-end
-
-@static if VERSION >= v"1.12"
-    function (@main)(ARGS)
-        if length(ARGS) == 1 && ARGS[1] in ("-h", "--help")
-            return print_help()
-        end
-        return run_cli(ARGS)
-    end
-end
-
-end
