+++ src/KomaMRI.jl
@@ -25,6 +25,9 @@ include("ui/ExportUIFunctions.jl")
 include("callbacks/ui_progress_callback.jl")
 include("KomaUI.jl")
 include("KomaCLI.jl")
+@static if VERSION >= v"1.12"
+    include("KomaApp.jl")
+end
 
 # Export the UI and the observables
 export KomaUI
